src/main/java/com/example/ecommerce/controller/PriceController.java
package com.example.ecommerce.controller;

import com.example.ecommerce.entity.Price;
import com.example.ecommerce.exception.DataNotFoundException;
import com.example.ecommerce.exception.InvalidRequestException;
import com.example.ecommerce.model.PriceResponse;
import com.example.ecommerce.repository.PriceRepository;
import com.example.ecommerce.service.PriceService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.time.LocalDateTime;

@RestController
@RequestMapping("/api/prices")
public class PriceController {

    private final PriceService priceService;

    @Autowired
    public PriceController(PriceService priceService) {
        this.priceService = priceService;
    }

    @GetMapping("/getPrice")
    public ResponseEntity<PriceResponse> getPrice(
            @RequestParam("applicationDate") @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime applicationDate,
            @RequestParam("productId") Long productId,
            @RequestParam("brandId") Long brandId
    ) {
        try {
            PriceResponse response = priceService.getPrice(applicationDate, productId, brandId);
            return new ResponseEntity<>(response, HttpStatus.OK);
        } catch (DataNotFoundException e) {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        } catch (InvalidRequestException e) {
            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
        } catch (Exception e) {
            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
        }
    }
}